requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
google-auth>=2.0.0
google-auth-oauthlib>=0.5.0
gspread>=5.0.0
requests>=2.28.0
PyJWT>=2.4.0